+++ packages/bluesky/lib/src/entities/embed_view_record_view_blocked.dart
@@ -8,6 +8,10 @@
 import 'package:atproto_core/atproto_core.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
 
+// 🌎 Project imports:
+import 'blocked_author.dart';
+import 'keys/ids.g.dart';
+
 part 'embed_view_record_view_blocked.freezed.dart';
 part 'embed_view_record_view_blocked.g.dart';
 
@@ -27,11 +31,18 @@ class EmbedViewRecordViewBlocked with _$EmbedViewRecordViewBlocked {
   /// identifier (URI) of the blocked record.
   const factory EmbedViewRecordViewBlocked({
     /// A string that represents the type of the blocked record view.
-    @typeKey required String type,
+    @typeKey @Default(appBskyEmbedRecordViewBlocked) String type,
 
     /// An `AtUri` instance that contains the unique resource identifier (URI)
     /// of the blocked record.
     @atUriConverter required AtUri uri,
+
+    /// Represents this view is blocked.
+    /// Always `true`.
+    @JsonKey(name: 'blocked') required bool isBlocked,
+
+    /// The author of this blocked view record.
+    required BlockedAuthor author,
   }) = _EmbedViewRecordViewBlocked;
 
   /// Creates an instance of [EmbedViewRecordViewBlocked] from a map of