+++ packages/bluesky/lib/src/entities/embed_view_record_view_not_found.g.dart
@@ -15,13 +15,15 @@ _$_EmbedViewRecordViewNotFound _$$_EmbedViewRecordViewNotFoundFromJson(
       json,
       ($checkedConvert) {
         final val = _$_EmbedViewRecordViewNotFound(
-          type: $checkedConvert(r'$type', (v) => v as String),
+          type: $checkedConvert(
+              r'$type', (v) => v as String? ?? appBskyEmbedRecordViewNotFound),
           uri: $checkedConvert(
               'uri', (v) => atUriConverter.fromJson(v as String)),
+          isNotFound: $checkedConvert('notFound', (v) => v as bool),
         );
         return val;
       },
-      fieldKeyMap: const {'type': r'$type'},
+      fieldKeyMap: const {'type': r'$type', 'isNotFound': 'notFound'},
     );
 
 Map<String, dynamic> _$$_EmbedViewRecordViewNotFoundToJson(
@@ -29,4 +31,5 @@ Map<String, dynamic> _$$_EmbedViewRecordViewNotFoundToJson(
     <String, dynamic>{
       r'$type': instance.type,
       'uri': atUriConverter.toJson(instance.uri),
+      'notFound': instance.isNotFound,
     };